Move signature code to validation module
